requirements.txt
streamlit>=1.36
ultralytics>=8.2
opencv-python>=4.8
pillow>=10.0
numpy>=1.24
